Fix compile warnings (bug #389)
git-svn-id: https://origsvn.digium.com/svn/asterisk/trunk@1644 65c4cc65-6c06-0410-ace0-fbb531ad65f3

+++ channels/chan_iax.c
@@ -533,7 +533,7 @@ static void showframe(struct ast_iax_frame *f, struct ast_iax_full_hdr *fhi, int
 	retries, ntohs(fh->seqno), class, subclass);
 		fprintf(stderr,
 "   Timestamp: %05ldms  Callno: %5.5d  DCall: %5.5d [%s:%d]\n",
-	ntohl(fh->ts),
+	(long)ntohl(fh->ts),
 	(short)(ntohs(fh->callno) & ~AST_FLAG_FULL), (short) ntohs(fh->dcallno),
 		inet_ntoa(sin->sin_addr), ntohs(sin->sin_port));
 }
