+++ channels/iax2-parser.c
@@ -64,7 +64,7 @@ static void dump_string(char *output, int maxlen, void *value, int len)
 static void dump_int(char *output, int maxlen, void *value, int len)
 {
 	if (len == sizeof(unsigned int))
-		snprintf(output, maxlen, "%ld", ntohl(*((unsigned int *)value)));
+		snprintf(output, maxlen, "%ld", (long)ntohl(*((unsigned int *)value)));
 	else
 		snprintf(output, maxlen, "Invalid INT");
 }
@@ -288,7 +288,7 @@ snprintf(tmp, sizeof(tmp),
 	outputf(tmp);
 snprintf(tmp, sizeof(tmp), 
 "   Timestamp: %05ldms  SCall: %5.5d  DCall: %5.5d [%s:%d]\n",
-	ntohl(fh->ts),
+	(long)ntohl(fh->ts),
 	ntohs(fh->scallno) & ~IAX_FLAG_FULL, ntohs(fh->dcallno) & ~IAX_FLAG_RETRANS,
 		inet_ntoa(sin->sin_addr), ntohs(sin->sin_port));
 	outputf(tmp);
